# services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Ocserv User management Api server.",
        "title": "Ocserv User management Example Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/customers/disconnect_sessions": {
            "post": {
                "description": "disconnects all online sessions for a customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Disconnect all online sessions of a customer",
                "parameters": [
                    {
                        "description": "customer username and password (same ocserv account).",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/customer.SummaryData"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middlewares.TooManyRequests"
                        }
                    }
                }
            }
        },
        "/customers/summary": {
            "post": {
                "description": "Customer summary account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Customer summary account",
                "parameters": [
                    {
                        "description": "customer username and password (same ocserv account).",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/customer.SummaryData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/customer.SummaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middlewares.TooManyRequests"
                        }
                    }
                }
            }
        },
...
        "/occtl/commands": {
            "get": {
                "description": "Occtl Commands",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCCTL"
                ],
                "summary": "Occtl Commands",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Command Action ID (1 to 15)",
                        "name": "action",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional parameter depending on command",
                        "name": "value",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
        "/occtl/server_info": {
            "get": {
                "description": "Server information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCCTL"
                ],
                "summary": "Server information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ServerVersion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "customer.ModelCustomer": {
            "type": "object",
            "required": [
                "deactivated_at",
                "expire_at",
                "is_locked",
                "owner",
                "rx",
                "traffic_size",
                "traffic_type",
                "tx",
                "username"
            ],
            "properties": {
                "deactivated_at": {
                    "type": "string"
                },
                "expire_at": {
                    "type": "string"
                },
                "is_locked": {
                    "type": "boolean"
                },
                "owner": {
                    "type": "string"
                },
                "rx": {
                    "description": "Receive in bytes",
                    "type": "integer"
                },
                "traffic_size": {
                    "description": "in GiB  \u003e\u003e x * 1024 ** 3",
                    "type": "integer"
                },
                "traffic_type": {
                    "type": "string",
                    "enum": [
                        "Free",
                        "MonthlyTransmit",
                        "MonthlyReceive",
                        "TotallyTransmit",
                        "TotallyReceive"
                    ]
                },
                "tx": {
                    "description": "Transmit in bytes",
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "customer.SummaryData": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                }
            }
        },
        "customer.SummaryResponse": {
            "type": "object",
            "required": [
                "ocserv_user",
                "usage"
            ],
            "properties": {
                "ocserv_user": {
                    "$ref": "#/definitions/customer.ModelCustomer"
                },
                "usage": {
                    "$ref": "#/definitions/customer.UsageResponse"
                }
            }
        },
        "customer.UsageResponse": {
            "type": "object",
            "required": [
                "bandwidths",
                "date_end",
                "date_start"
            ],
            "properties": {
                "bandwidths": {
                    "$ref": "#/definitions/repository.TotalBandwidths"
                },
                "date_end": {
                    "type": "string"
                },
                "date_start": {
                    "type": "string"
                }
            }
        },
...
        "middlewares.TooManyRequests": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "middlewares.Unauthorized": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.ServerVersion": {
            "type": "object",
            "properties": {
                "occtl_version": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "repository.TotalBandwidths": {
            "type": "object",
            "required": [
                "rx",
                "tx"
            ],
            "properties": {
                "rx": {
                    "type": "number"
                },
                "tx": {
                    "type": "number"
                }
            }
        },
        "request.ErrorResponse": {
            "type": "object",
            "required": [
                "error",
                "message"
            ],
            "properties": {
                "error": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}